expanded lisp_value represenation
Before we were only concerned with the LSB(3) bits, which we will
now refer to as "small tags", but this limited us to only 8 distinct
tags and then we decided to maximize fixnum's width to 63 bits which
halved our 8 distinct small tags to 4! Now we have a small tag that
can give us another 5 bits to represent as "wide tags". This now
gives us 32 distinct wide tags and 56 bits worth of storage in a 64-
bit lisp_value.