implemented a CONS pool and HANDLER-CASE / SIGNAL
Interestingly, most CONS pairs are allocated where their CDR will
be the next CONS allocated, or in other words most CONS operations
are for constructing lists sequentially with few intermediate CONSes.
With some experimentation it was found that alloctinging conses with
a pool allocator where every cons is next to another cons had up to
a 14% decrease in execution time when operating on larger lists.
This is supposedly because most long lists are allocated in the form:

ADDR
0x1000: CAR -> val
0x1008: CDR -> 0x1010
0x1010: CAR -> val
0x1018: CDR -> 0x1020
0x1020: CAR -> val
0x1028: CDR -> 0x1030
0x1030: CAR -> val
0x1038: CDR -> ...

though it isn't space efficient, it should result in fewer CPU cache
misses.