Hacky re-implementation of LABELS and FLET
Prior to converting to being a Lisp-2, FLET and LABELS were fairly
straightforward to implement as just syntax sugar for LET and LET*
but after changing to a Lisp-2 it's no longer to call a variable.
The current implementation associates functions and symbols directly
which was approximately a 10x performance boost compared to the
Lisp-1 model that must lookup every function within the environment.

This meant that FLET and LABELS must be re-implemented. My first try
was to combine MAKE-SYMBOL and remap the user defined names for the
functions to their new not-interned symbols and then at runtime it
would %DEFINE-FUNCTION using the new symbol. This lead to a bug
where recursive calls to a function using one of these forms would
have the labeled function replaced with one in the current, new/
recrusive, scope. This means when the recursive function returns,
if it executes its labeled function, it will be executing in the.
context from when it was redefined. The WHILE and UNTIL forms were
implemented in this way as tail-recursive LABELS.

Now my second attempt is to again generate not-interned symbols and
transform the symbols in the call position into a respective FUNCALL
form, and to turn the (FUNCTION SYM) and #'SYM forms into just SYM.