new calling convention, doesn't allocate as much
New calling convention where instead of popping + consing the
arguments before the call, just pass a change the top of stack
pointer and pass that directly. This has zero allocation overhead
when calling primitive functions and now the only leak is when
calling lisp lambdas is binding to the environment.